+++ src/home/flibrary/logic/model/TreeModel.cpp
@@ -29,7 +29,7 @@ QModelIndex TreeModel::index(const int row, const int column, const QModelIndex 
 		return {};
 
 	const auto * parentItem = parent.isValid() ? static_cast<IDataItem *>(parent.internalPointer()) : m_data.get();
-	const auto childItem = parentItem->GetChild(row);
+	const auto childItem = parentItem->GetChild(static_cast<size_t>(row));
 
 	return childItem ? createIndex(row, column, childItem.get()) : QModelIndex();
 }
